Move email messages into separate module
